Move tellolib to own folder for clarity
