+++ dronelib/__init__.py
@@ -1,0 +1,1 @@
+# This is needed to import tellolib
