+++ dronelib/util.py
