--- land_tello.py
@@ -1,9 +1,0 @@
-from dronelib_tello import TelloDrone
-
-def main():
-    drone = TelloDrone()
-    drone.activate()
-    drone.drone.send_command("land")
-
-if __name__ == "__main__":
-        main()
