+++ tello_demo.py
@@ -1,13 +1,13 @@
-from dronelib_tello import TelloDrone
-from util import *
+from dronelib.dronelib_tello import TelloDrone
+from dronelib.util import *
 from time import sleep
 
 def main():
     drone = TelloDrone()
     drone.activate()
     drone.takeoff()
-    #image = drone.camera_image()
-    #save_image(image)
+    image = drone.camera_image()
+    save_image(image)
 
     drone.land()
